requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-Bcrypt==1.0.1
mysqlclient==2.2.0
